changed distance to double and removed debug

+++ api/src/main/java/de/funboyy/bingo/api/waypoint/WayPoint.java
@@ -20,8 +20,8 @@ public interface WayPoint extends WorldObject {
 
   void scale(final float scale);
 
-  void distance(final float distance);
+  void distance(final double distance);
 
-  float distance();
+  double distance();
 
 }
